pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "elephantbroker"
version = "0.1.0"
description = "Unified Cognitive Runtime for the OpenClaw platform (OMNIA Protocol)"
requires-python = ">=3.11"
license = "AGPL-3.0-only"
dependencies = [
    "pydantic>=2.0,<3.0",
    "cognee[neo4j]==0.5.3",
    "cognee-community-vector-adapter-qdrant>=0.2.2",
    "httpx>=0.27",
    "qdrant-client>=1.7",
    "redis>=5.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "opentelemetry-api>=1.24",
    "opentelemetry-sdk>=1.24",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "click>=8.1",
    "prometheus_client>=0.21",
    "pyyaml>=6.0",
    "clickhouse-connect>=0.7",
    "mistralai>=1.0",
]

[project.scripts]
elephantbroker = "elephantbroker.server:main"
ebrun = "elephantbroker.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.4",
    "tiktoken>=0.7",
    "websockets>=13.0",
]
scenario = [
    "websockets>=13.0",
    "PyNaCl>=1.5.0",
]

[tool.setuptools.packages.find]
include = ["elephantbroker*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
markers = [
    "integration: integration tests requiring live infrastructure (Neo4j, Qdrant, Redis)",
    "pipeline: pipeline integration tests requiring real LLM (slow)",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
